requirements.txt
click==6.7
requests==2.18.4
scout.py==0.1.5
pathlib2==2.3.0
